Adicionando a refatoração de código do CRUD de usuario e Login. Com Criptografia e o JWT.

+++ src/Controller/UsuarioController.ts
@@ -2,21 +2,32 @@ import Usuario from '../entities/model/Usuario';
 import { Request, Response } from 'express';
 
 import statusCode from '../config/statusCode';
-
-import UsuarioService from '../Service/UsuarioService';
+import validations from '../validations/Validations';
+import Criptografar from '../config/Criptografar';
 
 class UsuarioController {
 
     public async Cadastrar(req: Request, res: Response) {
-        const { nomeUsuario, email, senha } = req.body;
-        const usuario = new Usuario({ nomeUsuario: nomeUsuario, email: email, senha: senha });
+        try {
+            const { nome, nomeUsuario, email, senha } = req.body;
 
-        const result = await UsuarioService.cadastrar(usuario);
+            const validar = await validations.validarPreenchimentoCamposObrigatorios(nome, nomeUsuario, email, senha);
 
-        console.log(result.nomeUsuario);
+            if (validar != "") {
+                return res.status(statusCode.bad).json(validar);
+            }
+            if (await validations.verifcarExisteUsuarioByEmail(email) != null) {
+                return res.status(statusCode.success).json("Já existe um usuário com esse e-mail cadastrado. Tente com outro e-mail!");
+            } else {
+                const senhaCriptografada = await Criptografar.criptografar(senha);
+                const usuario = new Usuario({ nome: nome, nomeUsuario: nomeUsuario, email: email, senha: senhaCriptografada, papel: "usuario" });
+                await Usuario.create(usuario);
+                return res.status(statusCode.success).json(usuario);
+            }
 
-        return res.json(result);
-
+        } catch (error) {
+            return res.status(statusCode.error).send('Ocorreu um erro inesperado ao tentar cadastrar usuario!');
+        }
     }
 
     public async ListarTodos(req: Request, res: Response) {
@@ -37,13 +48,13 @@ class UsuarioController {
     public async BuscarPorId(req: Request, res: Response) {
         try {
             const { id } = req.params;
-            const usuario = await Usuario.findById(id);
 
-            if (usuario != null) {
-                return res.status(statusCode.success).json(usuario);
+            if (await validations.verifcarExisteUsuarioById(id) == null) {
+                return res.status(statusCode.not_found).send('Não existe usuario cadastrado com esse Id!');
             }
             else {
-                return res.status(statusCode.not_found).send('Não tem usuarios cadastrados!');
+                const usuario = await Usuario.findById(id);
+                return res.status(statusCode.success).json(usuario);
             }
         } catch (error) {
             return res.status(statusCode.error).send('Error Listen!');
@@ -53,14 +64,22 @@ class UsuarioController {
     public async Atualizar(req: Request, res: Response) {
         try {
             const { id } = req.params;
-            const { nomeUsuario, email, senha } = req.body;
-            const usuario = await Usuario.findByIdAndUpdate(id, { nomeUsuario: nomeUsuario, email: email, senha: senha });
+            const { nome, nomeUsuario, email, senha } = req.body;
 
-            if (usuario != null) {
-                return res.status(statusCode.success).send('Usuario Atualizado!');
+            const validar = await validations.validarPreenchimentoCamposObrigatorios(nome, nomeUsuario, email, senha);
+
+            if (validar != "") {
+                return res.status(statusCode.bad).json(validar);
             }
-            else {
-                return res.status(statusCode.success).send('Não tem usuarios cadastrados!');
+
+            const optUsuario = await validations.verifcarExisteUsuarioByEmail(email);
+
+            if (optUsuario != null && optUsuario.id != id) {
+                return res.status(statusCode.bad).json("O e-mail para qual você está tentando atualizar já pertence a outro usuário. Tente novamente com outro e-mail!");
+            } else {
+                const senhaCriptografada = await Criptografar.criptografar(senha);
+                const usuario = await Usuario.findByIdAndUpdate(id, { nome: nome, nomeUsuario: nomeUsuario, email: email, senha: senhaCriptografada.toString() });
+                return res.status(statusCode.success).json(usuario);
             }
         } catch (error) {
             return res.status(statusCode.error).send('Error Update!');
@@ -70,8 +89,16 @@ class UsuarioController {
     public async Deletar(req: Request, res: Response) {
         try {
             const { id } = req.params;
-            const usuario = await Usuario.findByIdAndDelete(id);
-            return res.status(statusCode.success).send('Usuario Deletado com sucesso!');
+
+            const validar = await validations.verifcarExisteUsuarioById(id);
+
+            if (validar != null) {
+                const usuario = await Usuario.findByIdAndDelete(id);
+                return res.status(statusCode.success).json(usuario);
+            } else {
+                return res.status(statusCode.success).json("Não existe usuario cadastrado com esse Id: " + id);
+            }
+
         } catch (error) {
             return res.status(statusCode.error).send('Error ao deletar usuario!');
         }
